+++ ipwhois/tests/test_rdap.py
@@ -15,11 +15,11 @@ log = logging.getLogger(__name__)
 
 class TestRDAP(TestCommon):
 
-    def test__RDAP(self):
+    def test_RDAP(self):
 
         self.assertRaises(NetError, RDAP, 'a')
 
-    def test__RDAPLookup(self):
+    def test_lookup(self):
 
         data_dir = path.dirname(__file__)